+++ app/helpers/folio/cstypo_helper.rb
@@ -3,21 +3,16 @@
 module Folio::CstypoHelper
   CSTYPO_REGEXP = /(?<![<\/])\b([szkvaiou%]\b) /i
 
-  def cstypo(string, replace_newlines_with_br: false)
+  def cstypo(string)
     if string.present? && string.is_a?(String)
       runner = string
 
-      if replace_newlines_with_br
-        runner = runner.gsub(/(\r\n|\n)/, "<br>")
-      end
-
       if I18n.locale == :cs
-        runner = runner.gsub(CSTYPO_REGEXP, '\1&nbsp;')
-                       .gsub(/(\d+)\s+(\d+)/, '\1&nbsp;\2')
-                       .gsub(/(\d+)\s+(Kč)/, '\1&nbsp;\2')
+        runner = runner.gsub(CSTYPO_REGEXP, '\1 ')
+                       .gsub(/(\d+)\s+(\d+|Kč)/, '\1 \2')
       end
 
-      runner.html_safe
+      runner
     else
       string
     end